Refactor: Api's now have seperate files with shared index
