+++ src/API/electionApi.js
@@ -1,0 +1,13 @@
+// src/API/election.js
+
+import api from './api';
+
+/**
+ * Start a new election
+ * @param {number} voterCount
+ * @returns {Promise<Object>}
+ */
+export const startElection = async (voterCount) => {
+    const response = await api.post('/election/start', { voterCount });
+    return response.data;
+};
