+++ src/screens/Party/PartyScreen.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { addParty } from '../../API/VotingAPI';
+import { addParty } from '../../API';
 import { toast } from 'react-hot-toast';
 import { FaArrowLeft } from 'react-icons/fa';
 import './PartyScreen.css';
